[iOS] Fix build when linked using pods
These errors were only seen if react-native-track-player was linked using pods. Since RNTP uses a bunch of stuff that is only available in iOS 10 or newer, and since the podspec platform setting is iOS 9, it cannot compile (even when the parent project deployment target/swift settings are set to higher versions). Changing this to 10 fixes the issue.