Add prioritizeOverSize setting for Android buffer
Default is `false` to maintain backward compatability. If you're
configuring minBuffer and maxBuffer, you probably want to set this to
`true`. Setting to `true` ensures that your `minBuffer` setting will
always be met when possible. When it is false, ExoPlayer has a separate
targetBufferSize that may trigger loading of the resource to pause even
if minBuffer is not met.