[js] Update hooks (#935)
* Add 'pollTrackPlayerStates' option to 'pollTrackPlayerStates' hook

The pollTrackPlayerStates did not update its returned `position` value in case the player was not playing or buffering. However, the name and the `interval` parameter suggest that the progress is constantly returned. It is needed for use cases in which e.g. the user jumps to a different position in the track (via TrackPlayer.seekTo) while the track is paused. It's also needed for the successful implementation of a slider which can be moved by the user while the player is paused.

* Export not yet exported hooks - as documented in `index.d.ts`

* Update useTrackPlayerProgress type definition, improve another type definition and format the hooks type definition section

* Update hook documentation according to useTrackPlayerProgress hook changes