Add tip about `waitForBuffer`  for network media
I ran into the issue where the audio would pause immediately after buffering a stream and would require a second `play()` call to start playing.  After searching the issues found that a lot of people #601 #594 #616 #978.  While there is some discussion if this should be enabled by default or perhaps there is a deep issue. However, until it is resolve, (for some of us at least) this seems to fix the issue and making a note in a docs would be handy.