Warn about potential crash in "Getting Started"
I encountered a crash recently in the latest `v2.2.0-rc3` related to calling a method on `TrackPlayer` before setup was complete and thought it should maybe be documented more clearly.  The details are in #1475 but the gist is I had a race condition between calling `getQueue` and `setupPlayer`, and that didn't used to crash the app but now it does and that seems to be intentional.

Not super sure on the wording here, but I thought it would make sense to flip the order of what's "first" since you always want to register the playback service, and then calling `setupPlayer` comes later.

Update docs/API.md

Update docs/API.md

Update docs/API.md