docs(player.md): updated documentation to help avoid audio word drops at fast rates
The default AudioItem in SwiftAudioEx drops you down to the low-quality, zero-latency pitch
algorithm. But at high speeds (e.g. beyond 2x), it ends up dropping words from audio all the time.
This updates the documentation to recommend a tweak for people that avoids this problem. Note that
the ideal case would be to actually change the default pitch algorithm to avoid this problem.

fix #1837