fix: #1117 - safeguard reset() and add() against races
The bug describes this further, there are race conditions when you call
reset() or add() and expect the native side (AVPlayer) to actually have
been reset or added to correctly by the time these methods resolve.

The issue comes in SwiftAudioEx's design. You could ask why we're not
fixing these problems there. Two things: a) I'm already trying to with
https://github.com/doublesymmetry/SwiftAudioEx/pull/43, and b) there
are structural design problems in SwiftAudioEx that don't make the
ultimate resolution of this issue easy/obvious. It comes down to the
way that `add` and `reset` are decoupled from the actual AVPlayer
updates through onCurrentItemChanged in QueuedAudioPlayer; because of
this decoupling, it doesn't even allow you to hold on to a promise until
the AVPlayer has actually been updated prior to resolving it.

So in the meantime, this fallback. We basically poll the native side
until it's in the state we expect it to be in. Far from ideal, and I'm
super-open to other suggestions. But at the moment, if you don't patch
this, the following scenarios won't work correctly:

1. Calling `reset()` followed by `add()`. If the newly added track is
shorter than the previously active track in the queue, you'll
immediately get an erroneous PlaybackQueueEnded.
2. Calling `reset()` followed by `getQueue()`. This will yield a queue
of nonzero length because reset resolves even before the native side
of the queue in SwiftAudioEx has fully updated.
2. Calling `add()` (at the top position) followed by `getDuration()`,
`seek()`, or any other function that ultimately tunnels through to
AVPlayer. These operations will instead act on the previous top track.

This PR fixes all those issues by polling `duration`, which is one of
the most direct ways to tunnel all the way through to AVPlayer's
perception of the current track. This has two weaknesses:

* Polling. Le sigh.
* It won't be able to help you in cases where you add a track of the
exact same duration as the previous track you're replacing.

Note also that this PR does not fix several scenarios:
* `remove()`. This can't be fixed until
https://github.com/doublesymmetry/SwiftAudioEx/pull/43 is merged and
taken up by this project. In the meantime, if you remove the
currently-active item from the queue, operations will still act on the
removed item. See the linked PR and bug for more details.
* A bunch of new operations that SwiftAudioEx's current codebase
exposes, like `moveItem`, `replaceItem`, and other operations that
suffer the exact same problem. When those are taken up into this
codebase, someone will need to come back in to patch those as well.